+++ src/components/dictionaries/cobuild/engine.ts
@@ -22,6 +22,7 @@ export default function search (
   text: string,
   config: AppConfig
 ): Promise<COBUILDSearchResult> {
+  text = encodeURIComponent(text)
   return fetchDirtyDOM('https://www.iciba.com/' + text)
     .then(doc => handleDOM(doc, config.dicts.all.cobuild.options))
     .catch(() => {