+++ src/components/dictionaries/google/engine.ts
@@ -1,10 +1,28 @@
 import { handleNoResult } from '../helpers'
 import { AppConfig } from '@/app-config'
 import { DictSearchResult } from '@/typings/server'
-import { isContainChinese, isContainEnglish } from '@/_helpers/lang-check'
+import { isContainChinese, isContainJapanese, isContainKorean } from '@/_helpers/lang-check'
 import { first } from '@/_helpers/promise-more'
 
-export type GoogleResult = string
+export interface GoogleResult {
+  searchText: {
+    text: string
+    audio?: string
+  }
+  trans: {
+    text: string
+    audio?: string
+  }
+}
+
+interface GoogleRawResult {
+  json: string
+  base?: string
+  tl?: string
+  tk1?: number
+  tk2?: number
+  text: string
+}
 
 type GoogleSearchResult = DictSearchResult<GoogleResult>
 
@@ -13,16 +31,10 @@ export default function search (
   config: AppConfig,
 ): Promise<GoogleSearchResult> {
 
-  const chCode = config.langCode === 'zh-TW' ? 'zh-TW' : 'zh-CN'
-  let sl = 'auto'
-  let tl = chCode
-  if (isContainChinese(text)) {
-    sl = chCode
-    tl = 'en'
-  } else if (isContainEnglish(text)) {
-    sl = 'en'
-    tl = chCode
-  }
+  const sl = 'auto'
+  const tl = !isContainChinese(text) || isContainJapanese(text) || isContainKorean(text)
+    ? config.langCode === 'zh-TW' ? 'zh-TW' : 'zh-CN'
+    : 'en'
 
   return first([
     fetchWithToken('https://translate.google.com', sl, tl, text),
@@ -32,42 +44,62 @@ export default function search (
   .then(handleText)
 }
 
-function fetchWithToken (base: string, sl: string, tl: string, text: string): Promise<string> {
+function fetchWithToken (base: string, sl: string, tl: string, text: string): Promise<GoogleRawResult> {
   return fetch(base)
     .then(r => r.text())
-    .then<Response>(body => {
+    .then<GoogleRawResult>(body => {
       const tkk = (body.match(/TKK=(.*?)\(\)\)'\);/) || [''])[0]
         .replace(/\\x([0-9A-Fa-f]{2})/g, '') // remove hex chars
         .match(/[+-]?\d+/g)
       if (tkk) {
-        const tk = getTK(text, Number(tkk[2]), (Number(tkk[0]) + Number(tkk[1])))
+        const tk1 = Number(tkk[2])
+        const tk2 = Number(tkk[0]) + Number(tkk[1])
+        const tk = getTK(text, tk1, tk2)
         if (tk) {
-          return fetch(`${base}/translate_a/single?client=t&sl=${sl}&tl=${tl}&q=${text}&tk=${tk}&hl=en&dt=at&dt=bd&dt=ex&dt=ld&dt=md&dt=qca&dt=rw&dt=rm&dt=ss&dt=t&ie=UTF-8&oe=UTF-8&otf=1&ssel=0&tsel=0&kc=5`)
+          const encodedText = encodeURIComponent(text)
+          return fetch(`${base}/translate_a/single?client=t&sl=${sl}&tl=${tl}&q=${encodedText}&tk=${tk}&hl=en&dt=at&dt=bd&dt=ex&dt=ld&dt=md&dt=qca&dt=rw&dt=rm&dt=ss&dt=t&ie=UTF-8&oe=UTF-8&otf=1&ssel=0&tsel=0&kc=5`)
+            .then(r => r.text())
+            .then(json => ({ json, base, tl, tk1, tk2, text }))
         }
       }
       return handleNoResult()
     })
-    .then(r => r.text())
 }
 
-function fetchWithoutToken (sl: string, tl: string, text: string): Promise<string> {
-  return fetch(`https://translate.googleapis.com/translate_a/single?client=gtx&sl=${sl}&tl=${tl}&dt=t&q=${text}`)
+function fetchWithoutToken (sl: string, tl: string, text: string): Promise<GoogleRawResult> {
+  return fetch(`https://translate.googleapis.com/translate_a/single?client=gtx&sl=${sl}&tl=${tl}&dt=t&q=${encodeURIComponent(text)}`)
     .then(r => r.text())
+    .then(json => ({ json, text }))
 }
 
 function handleText (
-  text: string
+  { json, base, tl, tk1, tk2, text }: GoogleRawResult
 ): GoogleSearchResult | Promise<GoogleSearchResult> {
-  const json = JSON.parse(text.replace(/,+/g, ','))
+  const data = JSON.parse(json.replace(/,+/g, ','))
 
-  if (!json[0] || json[0].length <= 0) {
+  if (!data[0] || data[0].length <= 0) {
     return handleNoResult()
   }
 
-  const result: string = json[0].map(item => item[0]).join(' ')
+  const transText: string = data[0].map(item => item[0]).join(' ')
 
-  if (result.length > 0) {
-    return { result }
+  if (transText.length > 0) {
+    return {
+      result: {
+        trans: {
+          text: transText,
+          audio: tk1 || tk2
+            ? `${base}/translate_tts?ie=UTF-8&q=${encodeURIComponent(transText)}&tl=${tl}&total=1&idx=0&tk=${getTK(transText, tk1, tk2)}&client=t`
+            : ''
+        },
+        searchText: {
+          text,
+          audio: tk1 || tk2
+            ? `${base}/translate_tts?ie=UTF-8&q=${encodeURIComponent(text)}&tl=${data[2]}&total=1&idx=0&tk=${getTK(text, tk1, tk2)}&client=t`
+            : ''
+        }
+      }
+    }
   }
 
   return handleNoResult()