+++ test/specs/components/dictionaries/google/engine.spec.ts
@@ -11,7 +11,7 @@ const fetchbak = window.fetch
 describe('Dict/Google/engine', () => {
   beforeAll(() => {
     window.fetch = jest.fn((url: string) => Promise.resolve({
-      text: () => url.includes('sl') ? translation : homepage
+      text: () => /translate_a|googleapis\.com/.test(url) ? translation : homepage
     }))
   })
 
@@ -23,7 +23,10 @@ describe('Dict/Google/engine', () => {
     return search('any', appConfigFactory())
       .then(searchResult => {
         expect(searchResult.audio).toBeUndefined()
-        expect(searchResult.result).toBe('不要温柔地进入那个晚安。 愤怒，对光明的消逝愤怒。')
+        expect(searchResult.result.trans.text).toBe('不要温柔地进入那个晚安。 愤怒，对光明的消逝愤怒。')
+        expect(typeof searchResult.result.trans.audio).toBe('string')
+        expect(typeof searchResult.result.searchText.text).toBe('string')
+        expect(typeof searchResult.result.searchText.audio).toBe('string')
       })
   })
 })